pyproject.toml
[tool.poetry]
name = "api-message-consumer"
version = "0.1.0"
description = ""
authors = ["lazzarusxd <joaolazarofera@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiosmtplib = "^3.0.2"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
aio-pika = "^9.5.3"
pydantic = {extras = ["email"], version = "^2.10.3"}
twilio = "^9.3.8"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
